Update links to point to temporary netlify URL
Until we update the DNS for contribute.cncf.io, the current links will not work (or may point to a page on GitHub) because it is currently being used to redirect to this repository. This is a temporary change until the DNS is updated, then we can go back and change these URLs to point to contribute.cncf.io again.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>