Fix link to the last modified commit page
Since the content in our website is generated from multiple
repositories, the github URLs for things like viewing the last commit
for a page, or editing a page, all need to calculate the proper github
URL based on the current page's path.

This fixes the calculation for the last commit for a page, which is
displayed at the bottom of the page.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>